# Playwright Workspaces/apiDefinition.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playwright Workspaces",
        "description": "Full browser automation powered by Azure Playwright Workspaces remote browsers. Create sessions, navigate, click, type, scrape data, take screenshots, and automate multi-step web workflows — all running on managed cloud Chromium instances.",
        "version": "1.0.0",
        "contact": {
            "name": "Troy Taylor",
            "email": "troy@troystaylor.com"
        }
    },
    "host": "pw-workspaces-broker.azurewebsites.net",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/sessions/{sessionId}": {
            "get": {
                "operationId": "GetSession",
                "summary": "Get Session Status",
                "description": "Get the current status of a browser session including page URL, title, and time remaining.",
                "tags": [
                    "Sessions"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details",
                        "schema": {
                            "$ref": "#/definitions/SessionStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found or expired"
                    }
                }
            },
            "delete": {
                "operationId": "CloseSession",
                "summary": "Close Session",
                "description": "Close a browser session and release the remote browser. Always close sessions when done to avoid unnecessary charges.",
                "tags": [
                    "Sessions"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session closed"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
...
        "/sessions/{sessionId}/content": {
            "get": {
                "operationId": "GetPageContent",
                "summary": "Get Page Content",
                "description": "Get the text content or HTML of the current page.",
                "tags": [
                    "Extraction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "type": "string",
                        "enum": [
                            "text",
                            "html",
                            "markdown"
                        ],
                        "description": "Output format (default: text)",
                        "x-ms-summary": "Format"
                    },
                    {
                        "name": "selector",
                        "in": "query",
                        "type": "string",
                        "description": "CSS selector to extract content from a specific element (optional — defaults to full page)",
                        "x-ms-summary": "Selector"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page content",
                        "schema": {
                            "$ref": "#/definitions/ContentResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/element-text": {
            "post": {
                "operationId": "GetElementText",
                "summary": "Get Element Text",
                "description": "Get the text content of a specific element by CSS selector.",
                "tags": [
                    "Extraction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "selector"
                            ],
                            "properties": {
                                "selector": {
                                    "type": "string",
                                    "description": "CSS selector of the element",
                                    "x-ms-summary": "Selector"
                                },
                                "all": {
                                    "type": "boolean",
                                    "description": "Get text from all matching elements (default: false, returns first match)",
                                    "x-ms-summary": "All Matches"
                                },
                                "attribute": {
                                    "type": "string",
                                    "description": "Get an attribute value instead of text (e.g., href, src)",
                                    "x-ms-summary": "Attribute"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Element text",
                        "schema": {
                            "$ref": "#/definitions/ElementTextResponse"
                        }
                    },
                    "404": {
                        "description": "Session or element not found"
                    }
                }
            }
        },
        "/sessions/{sessionId}/scrape": {
            "post": {
                "operationId": "ScrapeData",
                "summary": "Scrape Structured Data",
                "description": "Extract structured data from the page using CSS selectors. Define a mapping of field names to selectors and get back clean JSON.",
                "tags": [
                    "Extraction"
                ],
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "Session ID",
                        "x-ms-summary": "Session ID",
                        "x-ms-url-encoding": "single"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "fields"
                            ],
                            "properties": {
                                "containerSelector": {
                                    "type": "string",
                                    "description": "CSS selector for repeating item containers (e.g., '.product-card'). Omit for single-item extraction.",
                                    "x-ms-summary": "Container Selector"
                                },
                                "fields": {
                                    "type": "object",
                                    "description": "Map of field names to CSS selectors (e.g., {\"title\": \"h2.name\", \"price\": \".price\", \"link\": \"a@href\"}). Append @attr to get an attribute.",
                                    "x-ms-summary": "Field Selectors"
                                },
                                "limit": {
                                    "type": "integer",
                                    "description": "Maximum number of items to extract (default: 50)",
                                    "x-ms-summary": "Limit"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scraped data",
                        "schema": {
                            "$ref": "#/definitions/ScrapeResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "SessionStatusResponse": {
            "type": "object",
            "properties": {
                "sessionId": {
                    "type": "string",
                    "description": "Session identifier",
                    "x-ms-summary": "Session ID"
                },
                "url": {
                    "type": "string",
                    "description": "Current page URL",
                    "x-ms-summary": "URL"
                },
                "title": {
                    "type": "string",
                    "description": "Current page title",
                    "x-ms-summary": "Title"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Session creation time",
                    "x-ms-summary": "Created At"
                },
                "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Session expiration time",
                    "x-ms-summary": "Expires At"
                },
                "actionCount": {
                    "type": "integer",
                    "description": "Number of actions performed",
                    "x-ms-summary": "Action Count"
                }
            }
        },
...
        "ContentResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "description": "Current page URL",
                    "x-ms-summary": "URL"
                },
                "title": {
                    "type": "string",
                    "description": "Page title",
                    "x-ms-summary": "Title"
                },
                "content": {
                    "type": "string",
                    "description": "Page content in the requested format",
                    "x-ms-summary": "Content"
                },
                "format": {
                    "type": "string",
                    "description": "Content format (text, html, or markdown)",
                    "x-ms-summary": "Format"
                }
            }
        },
        "ElementTextResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string",
                    "description": "Text content of the element (first match)",
                    "x-ms-summary": "Text"
                },
                "texts": {
                    "type": "array",
                    "description": "Text content of all matching elements (when all=true)",
                    "items": {
                        "type": "string"
                    }
                },
                "count": {
                    "type": "integer",
                    "description": "Number of matching elements",
                    "x-ms-summary": "Match Count"
                }
            }
        },
        "ScrapeResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "description": "Extracted data items",
                    "items": {
                        "type": "object"
                    }
                },
                "count": {
                    "type": "integer",
                    "description": "Number of items extracted",
                    "x-ms-summary": "Count"
                },
                "url": {
                    "type": "string",
                    "description": "Page URL data was scraped from",
                    "x-ms-summary": "Source URL"
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-Key"
        }
    },
    "security": [
        {
            "api_key": []
        }
    ],
...
}